modified:   src/dsp/audio_func.h 	modified:   src/dsp/biquad.cpp 	modified:   src/dsp/biquad.h 	modified:   src/dsp/crossfade.cpp 	modified:   src/dsp/crossfade.h 	modified:   src/dsp/downmix.h 	modified:   src/dsp/dsp.cpp 	modified:   src/dsp/dsp.h 	modified:   src/dsp/eq.cpp 	modified:   src/dsp/eq.h 	modified:   src/dsp/fade.cpp 	modified:   src/dsp/fade.h 	modified:   src/dsp/gain.cpp 	modified:   src/dsp/gain.h 	modified:   src/dsp/hp.cpp 	modified:   src/dsp/hp.h 	modified:   src/dsp/lp.cpp 	modified:   src/dsp/lp.h 	modified:   src/dsp/lr_bal.cpp 	modified:   src/dsp/lr_bal.h 	modified:   src/stream.h
Change data buffers from int to double for DSP.

+++ src/dsp/crossfade.cpp
@@ -1,5 +1,5 @@
 #include "crossfade.h"
 
-XFade::XFade(int gain_db) {
-    
+XFade::XFade(double gain_db) {
+
 }
