+++ src/dsp/lr_bal.cpp
@@ -1,5 +1,5 @@
 #include "lr_bal.h"
 
-LR_Balance::LR_Balance(int angle) {
-    
+LR_Balance::LR_Balance(double angle) {
+
 }
